postgresql_query, postgresql_script: get rid of code duplication (#197)
